requirements.txt
cryptography==38.0.3
Django==3.2
django-cors-headers==3.13.0
djangorestframework==3.12.0
django-filter==22.1
drf-yasg2==1.19.4
flake8==4.*
Pillow==9.3.0
PyMySQL==1.0.2
pytest-cov==4.0.0
pytest-django==4.5.2
pytest-flake8==1.1.1
# uWSGI==2.0.21